Updated primitive directory structure and added test for vfs parent path
