+++ build_schema.sh
@@ -3,38 +3,8 @@
 set -e
 set -o pipefail
 
-for c in contracts/app/*; do
-    (cd $c && cargo schema)
-done
-
-# for c in contracts/data-storage/*; do
-#     (cd $c && cargo schema)
-# done
-
-for c in contracts/ecosystem/*; do
-    (cd $c && cargo schema)
-done
-
-for c in contracts/finance/*; do
-    (cd $c && cargo schema)
-done
-
-for c in contracts/fungible-tokens/*; do
-    (cd $c && cargo schema)
-done
-
-for c in contracts/modules/*; do
-    (cd $c && cargo schema)
-done
-
-for c in contracts/non-fungible-tokens/*; do
-    (cd $c && cargo schema)
-done
-
-for c in contracts/os/*; do
-    (cd $c && cargo schema)
-done
-
-for c in contracts/data-storage/*; do
-    (cd $c && cargo schema)
+for directory in contracts/*/; do
+    for contract in $directory/*/; do
+        ( cd $contract && cargo schema )
+    done
 done
