+++ contracts/data-storage/andromeda-primitive/src/contract.rs
@@ -1,23 +1,22 @@
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
-use cosmwasm_std::{ensure, Binary, Deps, DepsMut, Env, MessageInfo, Response, StdError, Storage};
+use cosmwasm_std::{ensure, Binary, Deps, DepsMut, Env, MessageInfo, Response};
 use cw2::{get_contract_version, set_contract_version};
 
-use crate::state::{
-    get_key_or_default, has_key_permission, query_value, DATA, KEY_OWNER, RESTRICTION,
-};
-use andromeda_data_storage::primitive::{
-    ExecuteMsg, InstantiateMsg, MigrateMsg, Primitive, PrimitiveRestriction, QueryMsg,
-};
+use andromeda_data_storage::primitive::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
 use andromeda_std::{
     ado_base::InstantiateMsg as BaseInstantiateMsg,
     ado_contract::ADOContract,
-    amp::AndrAddr,
     common::{context::ExecuteContext, encode_binary},
     error::{from_semver, ContractError},
 };
-use cw_utils::nonpayable;
 use semver::Version;
+
+use crate::{
+    execute::handle_execute,
+    query::{all_keys, get_value, owner_keys},
+    state::RESTRICTION,
+};
 
 // version info for migration info
 const CONTRACT_NAME: &str = "crates.io:andromeda-primitive";
@@ -64,85 +63,6 @@ pub fn execute(
     }
 }
 
-pub fn handle_execute(ctx: ExecuteContext, msg: ExecuteMsg) -> Result<Response, ContractError> {
-    match msg {
-        ExecuteMsg::SetValue { key, value } => execute_set_value(ctx.deps, ctx.info, key, value),
-        ExecuteMsg::DeleteValue { key } => execute_delete_value(ctx.deps, ctx.info, key),
-        ExecuteMsg::UpdateRestriction { restriction } => {
-            execute_update_restriction(ctx.deps, ctx.info, restriction)
-        }
-        _ => ADOContract::default().execute(ctx, msg),
-    }
-}
-
-pub fn execute_update_restriction(
-    deps: DepsMut,
-    info: MessageInfo,
-    restriction: PrimitiveRestriction,
-) -> Result<Response, ContractError> {
-    nonpayable(&info)?;
-    let sender = info.sender;
-    ensure!(
-        ADOContract::default().is_owner_or_operator(deps.storage, sender.as_ref())?,
-        ContractError::Unauthorized {}
-    );
-    RESTRICTION.save(deps.storage, &restriction)?;
-    Ok(Response::new()
-        .add_attribute("method", "update_restriction")
-        .add_attribute("sender", sender))
-}
-
-pub fn execute_set_value(
-    deps: DepsMut,
-    info: MessageInfo,
-    key: Option<String>,
-    value: Primitive,
-) -> Result<Response, ContractError> {
-    nonpayable(&info)?;
-    let sender = info.sender;
-    let key: &str = get_key_or_default(&key);
-    ensure!(
-        has_key_permission(deps.storage, &sender, key)?,
-        ContractError::Unauthorized {}
-    );
-    DATA.update::<_, StdError>(deps.storage, key, |old| match old {
-        Some(_) => Ok(value.clone()),
-        None => Ok(value.clone()),
-    })?;
-    // Update the owner of the key
-    KEY_OWNER.update::<_, StdError>(deps.storage, key, |old| match old {
-        Some(old) => Ok(old),
-        None => Ok(sender.clone()),
-    })?;
-
-    Ok(Response::new()
-        .add_attribute("method", "set_value")
-        .add_attribute("sender", sender)
-        .add_attribute("key", key)
-        .add_attribute("value", format!("{value:?}")))
-}
-
-pub fn execute_delete_value(
-    deps: DepsMut,
-    info: MessageInfo,
-    key: Option<String>,
-) -> Result<Response, ContractError> {
-    nonpayable(&info)?;
-    let sender = info.sender;
-
-    let key = get_key_or_default(&key);
-    ensure!(
-        has_key_permission(deps.storage, &sender, key)?,
-        ContractError::Unauthorized {}
-    );
-    DATA.remove(deps.storage, key);
-    KEY_OWNER.remove(deps.storage, key);
-    Ok(Response::new()
-        .add_attribute("method", "delete_value")
-        .add_attribute("sender", sender)
-        .add_attribute("key", key))
-}
-
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn migrate(deps: DepsMut, _env: Env, _msg: MigrateMsg) -> Result<Response, ContractError> {
     // New version
@@ -180,27 +100,9 @@ pub fn migrate(deps: DepsMut, _env: Env, _msg: MigrateMsg) -> Result<Response, C
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn query(deps: Deps, env: Env, msg: QueryMsg) -> Result<Binary, ContractError> {
     match msg {
-        QueryMsg::GetValue { key } => encode_binary(&query_value(deps.storage, key)?),
-        QueryMsg::AllKeys {} => encode_binary(&query_all_keys(deps.storage)?),
-        QueryMsg::OwnerKeys { owner } => encode_binary(&query_owner_keys(&deps, owner)?),
+        QueryMsg::GetValue { key } => encode_binary(&get_value(deps.storage, key)?),
+        QueryMsg::AllKeys {} => encode_binary(&all_keys(deps.storage)?),
+        QueryMsg::OwnerKeys { owner } => encode_binary(&owner_keys(&deps, owner)?),
         _ => ADOContract::default().query(deps, env, msg),
     }
-}
-
-fn query_all_keys(storage: &dyn Storage) -> Result<Vec<String>, ContractError> {
-    let keys = DATA
-        .keys(storage, None, None, cosmwasm_std::Order::Ascending)
-        .map(|key| key.unwrap())
-        .collect();
-    Ok(keys)
-}
-
-fn query_owner_keys(deps: &Deps, owner: AndrAddr) -> Result<Vec<String>, ContractError> {
-    let owner = owner.get_raw_address(deps)?;
-    let keys = KEY_OWNER
-        .range(deps.storage, None, None, cosmwasm_std::Order::Ascending)
-        .filter(|x| x.as_ref().unwrap().1 == owner)
-        .map(|key| key.unwrap().0)
-        .collect();
-    Ok(keys)
 }
