+++ contracts/data-storage/andromeda-primitive/src/lib.rs
@@ -1,6 +1,9 @@
 pub mod contract;
+mod execute;
 #[cfg(all(not(target_arch = "wasm32"), feature = "testing"))]
 pub mod mock;
-pub mod state;
+mod query;
+mod state;
+
 #[cfg(test)]
-mod tests;
+mod testing;
