+++ contracts/data-storage/andromeda-primitive/src/testing/mod.rs
@@ -1,0 +1,2 @@
+mod mock;
+mod tests;
